Functions that call cloudCompiler now must supply source code and open the compile modal window. Inverted logic in loadInto to handle bad data first.
